+++ route/src/main/java/ru/otus/route/service/PriceService.java
@@ -6,12 +6,12 @@ import ru.otus.common.model.TaxiType;
 @Service
 public class PriceService {
 
-    public static int STANDARD_TARIFF = 50;
-    public static float COMFORT_COEFF = 1.5F;
-    public static float BUSINESS_COEFF = 2.5F;
-    public static float PREMIUM_COEFF = 4.0F;
+    public static long STANDARD_TARIFF = 50;
+    public static double COMFORT_COEFF = 1.5;
+    public static double BUSINESS_COEFF = 2.5;
+    public static double PREMIUM_COEFF = 4.0;
 
-    public int getPriceOf1kmByTaxiType(TaxiType type) {
+    public Long getPriceOf1kmByTaxiType(TaxiType type) {
         switch (type) {
             case STANDARD: return STANDARD_TARIFF;
             case COMFORT: return Math.round(STANDARD_TARIFF * COMFORT_COEFF);
